Move configuration generation into separate module
seems more sensible and easier to digest this way

in the process, deduplicated `watchDirs`
