+++ lib/cli.js
@@ -1,6 +1,6 @@
 let { abort } = require("faucet-pipeline-core/lib/util");
 let parseArgs = require("minimist");
-let aiur = require(".");
+let generateConfig = require("./config");
 let path = require("path");
 
 let HELP = `
@@ -66,41 +66,3 @@ module.exports = function parseCLI(argv = process.argv.slice(2), help = HELP) {
 		options: { watch, fingerprint, sourcemaps, compact }
 	};
 };
-
-function generateConfig(source, target, referenceDir, assetsConfig) {
-	let componentsDir = "./components"; // TODO: customizable
-	let baseURI = "/"; // TODO: customizable
-
-	let { watchDirs, sass, plugins } = assetsConfig;
-
-	// merge `watchDirs`
-	let defaults = [source, componentsDir];
-	watchDirs = watchDirs ? defaults.concat(watchDirs) : defaults;
-
-	// merge `sass`
-	defaults = [{
-		source: "aiur/lib/style.scss", // TODO: customizable?
-		target: `${target}/style.css`
-	}];
-	sass = sass ? defaults.concat(sass) : defaults;
-
-	// merge `plugins`
-	defaults = {
-		aiur: {
-			plugin: aiur,
-			bucket: "markup"
-		}
-	};
-	plugins = plugins ? Object.assign({}, defaults, plugins) : defaults;
-
-	return Object.assign({}, assetsConfig, {
-		watchDirs,
-		aiur: [{ source, target, baseURI }],
-		sass,
-		manifest: { // TODO: customizable?
-			baseURI,
-			webRoot: target
-		},
-		plugins
-	});
-}
